+++ FluidSimulation/simulation_cubic.cpp
@@ -532,7 +532,7 @@ void SimulationCubic::Step_Time(int framenum, vector<WaterSource> &sources){
     //velocity-evolution
 	Apply_External_Forces();
 	Pour_Source(framenum, sources);
-	printf("particle num: %d\n", particles.size());
+	printf("particle num: %d\n", (int)particles.size());
 	t = omp_get_wtime(); printf("apply external forces&pour source time cost: %.2fs\n", (t - t0 + 0.0)); t0 = t;
 
 	Get_Particles_Velocity(particles, vx, vy, vz, mask);
